% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_merge.R
\name{deep_merge}
\alias{deep_merge}
\title{Deep merge two lists}
\usage{
deep_merge(obj1, obj2)
}
\arguments{
\item{obj1}{The first object}

\item{obj2}{The second object}
}
\value{
The merged list
}
\description{
This function will merge two lists recursively. If both lists
are named lists, the keys will be merged. If both lists are
lists, they will be appended. Otherwise, the second list will
override the first list.
}
\examples{
deep_merge(list(a = 1, b = 2), list(b = 3, c = 4))
# $a
# [1] 1
#
# $b
# [1] 3
#
# $c
# [1] 4
}
