% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list_a_dataframe.R
\name{is_list_a_dataframe}
\alias{is_list_a_dataframe}
\title{Check if this list could be a data frame}
\usage{
is_list_a_dataframe(li)
}
\arguments{
\item{li}{list}
}
\value{
whether the list could be a data frame
}
\description{
Check if this list could be a data frame
}
\examples{
df <- list(
  a = c(1, 2, 3),
  b = c("a", "b", "c")
)
is_list_a_dataframe(df)
}
