% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_named_list.R
\name{is_named_list}
\alias{is_named_list}
\title{Check if an object is a named list}
\usage{
is_named_list(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
TRUE if the object is a named list, FALSE otherwise
}
\description{
Check if an object is a named list
}
\examples{
is_named_list(list(a = 1, b = 2)) # TRUE
is_named_list(list(1, 2)) # FALSE
is_named_list(NULL) # TRUE
is_named_list(list()) # TRUE
is_named_list(c(1, 2, 3)) # FALSE
}
