% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_as_data_frame.R
\name{list_as_data_frame}
\alias{list_as_data_frame}
\title{Convert a list to a data frame}
\usage{
list_as_data_frame(li)
}
\arguments{
\item{li}{A list}
}
\value{
A data frame
}
\description{
This function converts a list to a data frame. It is a wrapper around
\code{\link{as.data.frame}} that only converts atomic vectors.
}
\examples{
li <- list(
  a = c(1, 2, 3),
  b = c("a", "b", "c")
)
}
