% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_as_tibble.R
\name{list_as_tibble}
\alias{list_as_tibble}
\title{Turn a list into a dataframe}
\usage{
list_as_tibble(li)
}
\arguments{
\item{li}{A list}
}
\value{
A tibble
}
\description{
This will remove any non-atomic elements from the list and turn it into a dataframe.
}
\details{
Note: this will only work if the list of atomic vectors are
of the same length or length 1.
}
\examples{
li <- list(
  a = c(1, 2, 3),
  b = c("a", "b", "c")
)
list_as_tibble(li)
}
