% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_margin.R
\name{strip_margin}
\alias{strip_margin}
\title{Strip margin from a string}
\usage{
strip_margin(text, symbol = "\\\\|")
}
\arguments{
\item{text}{A character vector.}

\item{symbol}{The margin symbol to strip.}
}
\value{
A character vector with the margin stripped.
}
\description{
Strip margin from a string
}
\examples{
strip_margin(
  "hello_world:
  |  this_is: \"a yaml\""
)
}
