% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_object.R
\name{validate_object}
\alias{validate_object}
\title{Validate an object against a schema}
\usage{
validate_object(
  obj,
  what = c("api_component_spec", "api_file_format", "task_config", "task_control_method",
    "task_method", "task_metric"),
  obj_source = NULL,
  engine = c("ajv", "imjv"),
  error = TRUE
)
}
\arguments{
\item{obj}{Object to validate}

\item{what}{Type of schema to validate against}

\item{obj_source}{Source of the object (e.g. the path to the config)}

\item{engine}{Validation engine to use. Currently supported are "imjv" and "ajv".}

\item{error}{Whether to throw an error if validation fails}
}
\description{
Validate an object against a schema
}
\examples{
task_config <- list(
  viash_version = "0.8.0",
  name = "task_example",
  organization = "openproblems",
  license = "MIT",
  label = "Example task",
  keywords = c("example", "task"),
  summary = "An example task",
  description = "A longer description of the example task",
  authors = list(),
  links = list(
    repository = "foo"
  ),
  info = list(
    image = "image.svg"
  )
)
validate_object(task_config, what = "task_config")
}
